pyproject.toml
[project]
name = "mcp-workshop-ms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.60.0",
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.116.1",
    "fastmcp>=2.11.0",
    "feedparser>=6.0.11",
    "lxml>=6.0.0",
    "mcp>=1.12.3",
    "openai>=1.98.0",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.0.0",
    "requests>=2.32.4",
]
